requirements.txt
streamlit
dotenv
pypdf2
google.generativeai
python-dotenv
python-docx